+++ bin/Print.lua
@@ -1,5 +1,4 @@
 local addon, tbl = ...
-local L = tbl.L
 
 tbl.Print = function(id, itemLink, count, source, location, lootDate, class, level)
 	if class and level then