+++ bin/Search_Text.lua
@@ -1,10 +1,9 @@
 local addon, tbl = ...
-local L = tbl.L
 
 tbl.Search_Text = function(query)
 	local itemsFound = 0
 	for k, v in pairs(tbl.Items) do
-		if v.source ~= L["INFO_MSG_MISCELLANEOUS"] or v.source or v.location or v.itemLink or v.lootDate then
+		if v.source ~= tbl.L["INFO_MSG_MISCELLANEOUS"] or v.source or v.location or v.itemLink or v.lootDate then
 			if string.find(string.lower(v.itemLink), string.lower(query)) or string.find(string.lower(v.itemName), string.lower(query)) then
 				local itemID = (GetItemInfoInstant(k))
 				if v.lootedBy["playerClass"] ~= nil then
@@ -44,8 +43,8 @@ tbl.Search_Text = function(query)
 		end
 	end
 	if itemsFound == 0 then
-		print(L["ADDON_NAME"] .. L["ERROR_MSG_NO_ITEMS_FOUND"] .. " (" .. query .. ")")
+		print(tbl.L["ADDON_NAME"] .. tbl.L["NO_ITEMS_FOUND"] .. " (" .. query .. ")")
 	else
-		print(L["ADDON_NAME"] .. itemsFound .. L["INFO_MSG_RESULTS"])
+		print(tbl.L["ADDON_NAME"] .. itemsFound .. tbl.L["RESULTS"])
 	end
 end